refactor: usable widget to its own file
